XRDDEV-2211 generating private parameters

+++ src/central-server/admin-service/globalconf-generator/src/main/java/org/niis/xroad/cs/admin/globalconf/generator/PrivateParameters.java
@@ -38,7 +38,7 @@ class PrivateParameters {
     private String instanceIdentifier;
     private List<ConfigurationAnchor> configurationAnchors;
     private ManagementService managementService;
-    private Long timeStampingIntervalSeconds;
+    private Integer timeStampingIntervalSeconds;
 
     @Data
     public static class ConfigurationAnchor {
